Scripting edits and some error fixes. see comments
current bit error is we need a different instance of the stack where the size is larger for the war stack. this should be 52 just in case, but logically it wont get to be larger than 6 (since we aren't doing double war). so we need one instance for WarStack that's size 6 and one for SidePile that is size 5. I also fixed the operator overloaders in list.cpp.